+++ Code/Common/UIBindingConstants.cs
@@ -17,6 +17,7 @@ namespace Planboard.Common
         public const string UndoAvailable = "undoAvailable";
         public const string WindowLayoutRevision = "windowLayoutRevision";
         public const string DeadlineMode = "deadlineMode";
+        public const string DataReadOnly = "dataReadOnly";
         public const string CreateEntry = "createEntry";
         public const string CreatePinnedDraft = "createPinnedDraft";
         public const string FinishDraft = "finishDraft";